+++ cypress/integration/performance/common/checkPageLoadTime.js
@@ -1,4 +1,4 @@
-const checkPageLoadTime = (maxTime) => {
+const checkPageLoadTime = (performance, maxTime) => {
     performance.measure("pageLoad", "start-loading", "end-loading");
     const measure = performance.getEntriesByName("pageLoad")[0];
     const duration = measure.duration;
